Move all API routes from /api to /app/api (App Router format)
Vercel with Next.js 16 requires API routes inside app/api/ as route.ts
files. Moved all endpoints and fixed import paths. Removed old /api folder.

Build confirms all routes are registered:
- /api/scan, /api/scan-once, /api/subscribers
- /api/admin/subscribers, /api/admin/scans, /api/admin/audit
- /api/admin/false-positives, /api/admin/export
- /api/auth/[...nextauth]
- /admin/* pages
