+++ app/api/admin/audit/route.ts
@@ -1,7 +1,7 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { auth } from "../../src/auth";
-import { getAuditLog, countAuditLog } from "../../src/audit-log";
+import { auth } from "../../../../src/auth";
+import { getAuditLog, countAuditLog } from "../../../../src/audit-log";
 
 export const runtime = "nodejs";
