+++ app/api/admin/export/route.ts
@@ -1,9 +1,9 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { auth } from "../../src/auth";
-import { getScanReport } from "../../src/scan-store";
-import { renderReportHtml } from "../../src/reporter";
-import { logAudit } from "../../src/audit-log";
+import { auth } from "../../../../src/auth";
+import { getScanReport } from "../../../../src/scan-store";
+import { renderReportHtml } from "../../../../src/reporter";
+import { logAudit } from "../../../../src/audit-log";
 
 export const runtime = "nodejs";
