+++ vercel.json
@@ -4,13 +4,5 @@
       "path": "/api/scan",
       "schedule": "0 6 * * *"
     }
-  ],
-  "functions": {
-    "api/scan.ts": {
-      "maxDuration": 300
-    },
-    "api/scan-once.ts": {
-      "maxDuration": 120
-    }
-  }
+  ]
 }
